third_party/blink/web_tests/platform/fuchsia/fast/url/invalid-urls-utf8-expected.txt

This ensure we do not mess up the bytes when an URL is partially parsed and is invalid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL src should be http:///. Was http:.
FAIL src should be https:///. Was https:.
FAIL src should be ftp:///. Was ftp:.
PASS src is expected
FAIL src should be ws:///. Was ws:.
FAIL src should be wss:///. Was wss:.
FAIL src should be foo://tête à tête@host/. Was foo://t%C3%AAte %C3%A0 t%C3%AAte@host/.
FAIL src should be foo://user:tête à tête@host/. Was foo://user:t%C3%AAte %C3%A0 t%C3%AAte@host/.
PASS src is expected
PASS src is expected
FAIL src should be foo://user:password@[xxxxx]/tête à tête/. Was foo://user:password@[xxxxx]/t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@[?tête à tête/. Was foo://user:password@[?t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@[?tête à tête]/. Was foo://user:password@[?t%C3%AAte %C3%A0 t%C3%AAte]/.
FAIL src should be foo://user:password@host:tête à tête/. Was foo://user:password@host:t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@host:80tête à tête. Was foo://user:password@host:80t%C3%AAte %C3%A0 t%C3%AAte.
FAIL src should be foo://user:password@host:tête à tête/. Was foo://user:password@host:t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@host:80tête à tête. Was foo://user:password@host:80t%C3%AAte %C3%A0 t%C3%AAte.
FAIL src should be ws:/tête. Was ws://xn--tte-fma/.
FAIL src should be ftp:/tête. Was ftp://xn--tte-fma/.
FAIL src should be http:/tête. Was http://127.0.0.1:8000/t%C3%AAte.
FAIL src should be https:/tête. Was https://xn--tte-fma/.
FAIL src should be gopher:/tête. Was gopher:/t%C3%AAte.
PASS successfullyParsed is true

TEST COMPLETE
